fix warning for C4389

+++ include/aquarius/server/io_service_pool.hpp
@@ -56,7 +56,7 @@ namespace aquarius
 				boost::asio::io_service& io_service = *io_services_[next_to_service_];
 				++next_to_service_;
 
-				if (next_to_service_ == io_services_.size())
+				if (next_to_service_ == static_cast<int>(io_services_.size()))
 					next_to_service_ = 0;
 
 				return io_service;
